feat: use macro to reduce duplicate ffi db code
